fix manifest generation and git push in renovate pipeline

+++ manifest-generate.sh
@@ -1,10 +1,11 @@
-#!/usr/bin/bash
+#!/bin/bash
 
 set -euo pipefail
 
+app_filter="${1:-}"
 generated_dir="generated"
 base_dir="$(pwd)"
-declare -a envs=("cl-infrasam-prod" "cl-infrasam-test")
+declare -a envs=("cl-infrasam-prod")
 echo "Base directory: $base_dir"
 
 cleanup() {
@@ -13,31 +14,40 @@ cleanup() {
 
 trap cleanup EXIT
 
-rm -rf "$generated_dir"
+if [[ -n "$app_filter" ]]; then
+  for env_dir in "${envs[@]}"; do
+    rm -rf "${generated_dir}/${env_dir}/${app_filter}"
+  done
+else
+  rm -rf "$generated_dir"
+fi
 
 for release_path in releases/*; do
-  for env_dir in "${envs[@]}"; do
-    if [[ -d "$release_path" ]]; then
-      release_name="${release_path##*/}"
-      if [[ "$release_name" == "template" ]]; then
-        continue
-      fi
+  if [[ -d "$release_path" ]]; then
+    release_name="${release_path##*/}"
+    if [[ "$release_name" == "template" ]]; then
+      continue
+    fi
+    if [[ -n "$app_filter" && "$release_name" != "$app_filter" ]]; then
+      continue
+    fi
 
+    for env_dir in "${envs[@]}"; do
       echo "Generating manifests for $release_name in $env_dir"
 
       output_dir="$base_dir/$generated_dir/$env_dir/$release_name"
       mkdir -p "$output_dir"
 
       pushd "$release_path" >/dev/null
-      helmfile -e "$(echo "$env_dir")" template --skip-tests --include-crds --log-level error >"$base_dir/$generated_dir/$env_dir/$release_name/manifest.lock.yaml"
+      helmfile -e "$env_dir" template --skip-tests --include-crds --log-level error >"$base_dir/$generated_dir/$env_dir/$release_name/manifest.lock.yaml"
       manifest="$base_dir/$generated_dir/$env_dir/$release_name/manifest.lock.yaml"
       if [ -f "$manifest" ] && [ ! -s "$manifest" ]; then
         echo "Manifest for $release_name in $env_dir is empty. Removing it."
         rm "$manifest"
       fi
       popd >/dev/null
-    fi
-  done
+    done
+  fi
 done
 
 echo "Combined YAML files written to $generated_dir"
